requirements.txt
h5py==3.8.0
matplotlib==3.7.0
opencv_python==4.7.0.72
Pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
trimesh==3.21.5
webdataset==0.2.48